script/requirements.txt
docx2txt
PyPDF2
Pillow
PyQt5
flask